app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
paho-mqtt==2.1.0
pyserial==3.5
numpy==1.26.4
tflite-runtime==2.14.0
httpx==0.27.2
